Submodule goflymake so we can run in any GOPATH
Previously, we found our goflymake lisp files using the $GOPATH
variable. This meant emacs had to be started with the GOPATH variable
set to whatever directory it was set to when `go get
github.com/dougm/goflymake` was run.

Now we submodule the goflymake lisp code we need into this repo. This
means it is possible to have a "global $GOPATH" where tools like
goflymake and ginkgo are compiled, and separately to have
project-specific $GOPATHs which are not poluted with goflymake.