Stop defaulting to windows-style shortcuts
While this started as an attempt to make pairing code-interviews
easier it's become my standard emacs config, and I hate not being able
to do a rectangle transform ( C-x r t ) because C-x is bound to
`kill-region` for windows consistenccy.

On the off-chance that any non-emacs people actually use this, they'll
probably be on a mac anyway.